// src/api/types.gen.ts
export type AddEventEntry = {
    name: string;
    eventDateStart: string;
    description?: string | null;
    eventDateEnd?: string | null;
};

export type AddPurchaseEntry = {
    name: string;
    cost: number;
    participantId: string;
    purchaseUsages: Array<UpsertPurchaseUsageEntry>;
};

export type ApiError = {
    errorCode?: string | null;
    message?: string | null;
    messageTemplate?: string | null;
    args?: Array<unknown> | null;
    httpStatus?: number;
};

export type EventShortEntry = {
    id: string;
    name: string;
    description?: string | null;
    eventDateStart: string;
    eventDateEnd?: string | null;
    isCompleted: boolean;
    imageUri?: string | null;
};
// ...
export type UpdateEventEntry = {
    name: string;
    description?: string | null;
    eventDateStart: string;
    eventDateEnd?: string | null;
};